continued progress on l2r-only test definitions
