+++ core/services/mongodb-remote/src/main/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/DataSynchronizer.java
@@ -439,10 +439,10 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
   public void stop() {
     syncLock.lock();
     try {
+      instanceChangeStreamListener.stop();
       if (syncThread == null) {
         return;
       }
-      instanceChangeStreamListener.stop();
       syncThread.interrupt();
       try {
         syncThread.join();
